requirements.txt
fastapi
python-multipart
python-jose
cryptography
passlib
typing
sqlalchemy
# pyodbc
pymssql
pyyaml
uvicorn
requests
loguru